No single letter variables
This allows `_` to be used as a variable name, since it is conventionally
used for throwaway variables, but does not allow any other single letter
variables.

The `target` method of `ast.Assigns` will normally contain a list of nodes,
however for an unpacking assignment (such as `'a,b = c'`) it will contain
an `ast.Tuple` or an `ast.List` so we need to cover for all cases.